+++ R/wf_get_key.R
@@ -2,23 +2,33 @@
 #'
 #' Returns you token set by \code{\link[ecmwfr]{wf_set_key}}
 #'
-#' @param email email address used to sign up for the ECMWF data service
+#' @param user user (email address) used to sign up for the ECMWF data service
+#' @param service which service to use
 #' @return the key set using \code{\link[ecmwfr]{wf_set_key}} saved
 #' in the keychain
 #' @keywords key management
 #' @seealso \code{\link[ecmwfr]{wf_set_key}}
 #' @export
+#' @author Koen Kufkens
 #' @examples
 #'
 #' \dontrun{
 #' # set key
-#' wf_set_key(email = "test@mail.com", key = "123")
+#' wf_set_key(user = "test@mail.com", key = "123")
 #'
 #' # get key
-#' wf_get_key(email = "test@mail.com")
+#' wf_get_key(user = "test@mail.com")
 #'}
 
-wf_get_key <- function(email){
-  keyring::key_get(service = "ecmwfr",
-                   username = email)
+wf_get_key <- function(
+  user,
+  service = "webapi"
+  ){
+  if(tolower(service) == "webapi"){
+    keyring::key_get(service = "ecmwfr_webapi",
+                     username = user)
+  }else{
+    keyring::key_get(service = "ecmwfr_cds",
+                     username = user)
+  }
 }