+++ R/wf_request.R
@@ -4,78 +4,104 @@
 #' you can only stage requests, logging the request URLs to submit download
 #' queries later on using \code{\link[ecmwfr]{wf_transfer}}. The function only
 #' allows NetCDF downloads, and will override calls for grib data.
+#' Note that the function will do some basic checks on the \code{request} input
+#' to identify possible problems.
 #'
-#' @param email email address used to sign up for the ECMWF data service and
+#' @param user user (email address) used to sign up for the ECMWF data service,
 #' used to retrieve the token set by \code{\link[ecmwfr]{wf_set_key}}
-#' @param path path where to store the downloaded data
-#' @param time_out how long to wait on a download to start (default = 3600)
+#' @param path path were to store the downloaded data
+#' @param time_out how long to wait on a download to start (default =
+#' \code{3*3600} seconds).
 #' @param transfer logical, download data TRUE or FALSE (default = FALSE)
 #' @param request nested list with query parameters following the layout
 #' as specified on the ECMWF API page
 #' @param verbose show feedback on processing
-#' @return a download query staging url or a netCDF of data on disk
+#' @return a download query staging url or (invisible) filename of the NetCDF
+#' file on your local disc
 #' @keywords data download, climate, re-analysis
 #' @seealso \code{\link[ecmwfr]{wf_set_key}}
 #' \code{\link[ecmwfr]{wf_transfer}}
 #' @export
+#' @author Koen Kufkens
 #' @examples
 #'
 #' \dontrun{
 #' # set key
-#' wf_set_key(email = "test@mail.com", key = "123")
+#' wf_set_key(user = "test@mail.com", key = "123")
+#'
+#' request <-  = list(stream = "oper",
+#'    levtype = "sfc",
+#'    param = "167.128",
+#'    dataset = "interim",
+#'    step = "0",
+#'    grid = "0.75/0.75",
+#'    time = "00",
+#'    date = "2014-07-01/to/2014-07-02",
+#'    type = "an",
+#'    class = "ei",
+#'    area = "50/10/51/11",
+#'    format = "netcdf",
+#'    target = "tmp.nc")
 #'
 #' # get the default test data
-#' wf_request(email = "test@mail.com")
+#' wf_request(user = "test@mail.com", request = request)
 #'}
 
 wf_request <- function(
   email,
-  path = tempdir(),
-  time_out = 3600,
+  service = "webapi",
+  request,
   transfer = FALSE,
-  request = list(stream = "oper",
-                 levtype = "sfc",
-                 param = "167.128",
-                 dataset = "interim",
-                 step = "0",
-                 grid = "0.75/0.75",
-                 time = "00",
-                 date = "2014-07-01/to/2014-07-02",
-                 type = "an",
-                 class = "ei",
-                 area = "50/10/51/11",
-                 format = "netcdf",
-                 target = "tmp.nc"),
+  path = tempdir(),
+  time_out = 3*3600,
   verbose = TRUE
   ){
 
+  # match arguments, if not stop
+  service <- match.arg(service, c("webapi", "cds"))
+
   # check the login credentials
-  if(missing(email)){
-    stop("Please provide ECMWF login credentials and data request!")
+  if(missing(email) || missing(request)){
+    stop("Please provide ECMWF or CDS login credentials and data request!")
   }
 
-  # get key from email
-  key <- wf_get_key(email)
+  # get key
+  key <- wf_get_key(email, service = service)
 
-  # force the use of netcdf
-  request$format <- "netcdf"
-  request$target <- paste0(tools::file_path_sans_ext(request$target),
-                           ".nc")
+  # getting api url: different handling if 'dataset = "mars"',
+  # requests to 'dataset = "mars"' require a non-public user
+  # account (member states/commercial).
+  url <- if(request$dataset == "mars") {
+    sprintf("%s/services/mars/requests", wf_server())
+  } else{
+    sprintf("%s/datasets/%s/requests", wf_server(), request$dataset)
+  }
 
-  # get the response from the query provided
-  response <- httr::POST(
-    paste(ecmwf_server(),
-          "datasets",
-          request$dataset,
-          "requests", sep = "/"),
-    httr::add_headers(
-      "Accept" = "application/json",
-      "Content-Type" = "application/json",
-      "From" = email,
-      "X-ECMWF-KEY" = key),
-    body = request,
-    encode = "json"
+  # depending on the service get the response
+  # for the query provided
+  if (service == "webapi"){
+    response <- httr::POST(
+      url,
+      httr::add_headers(
+        "Accept" = "application/json",
+        "Content-Type" = "application/json",
+        "From" = user,
+        "X-ECMWF-KEY" = key),
+      body = request,
+      encode = "json"
+      )
+  } else {
+    response <- httr::POST(
+      sprintf("%s/resources/%s", wf_server(service = "cds"),
+              request$dataset),
+      httr::authenticate(user, key),
+      httr::add_headers(
+        "Accept" = "application/json",
+        "Content-Type" = "application/json"),
+      body = request,
+      encode = "json"
     )
+  }
 
   # trap general http error
   if(httr::http_error(response)){
@@ -88,9 +114,17 @@ wf_request <- function(
 
   # some verbose feedback
   if(verbose){
-    message("Staging data transfer at url endpoint:")
-    message(ct$href)
+    message("- staging data transfer at url endpoint:")
+    message("  ", ct$href)
   }
+
+  # on exit show message
+  on.exit(
+    exit_message(
+      id = ct$request_id,
+      path = path,
+      file = request$target)
+    )
 
   # only return the content of the query
   if(!transfer){
@@ -98,24 +132,33 @@ wf_request <- function(
   }
 
   # set time-out counter
+  if(verbose) message(sprintf("- timeout set to %.1f hours", time_out/3600))
+
+  # set time-out
   time_out <- Sys.time() + time_out
 
+  # Temporary file name, will be used in combination with tempdir() when
+  # calling wf_transfer. The final file will be moved to the user-defined
+  # 'path' as soon as the download has been finished.
+  tmp_file <- basename(tempfile("ecmwfr_"))
+
   # keep waiting for the download order to come online
-  # with status code 303
+  # with status code 303. 202 = connection accepted, but job queued.
+  # http error codes (>400) will be trapped by the wf_transfer()
+  # function call
   while(ct$code == 202){
 
     # exit routine when the time out
-    # is reached, create message to consult
-    # the ecmwf website to download the data
-    # or retain the download url and use
-    # wf_transfer()
     if(Sys.time() > time_out){
-      message("Please use the MARS job list to track your jobs at:")
-      message("https://apps.ecmwf.int/webmars/joblist/")
-      message("and retry download using wf_transfer() for url:")
-      message(ct$href)
-      message("There is a limit of 3 active and 20 queued jobs.")
-      message("Delete the job using wf_delete() upon completion!")
+      if(verbose){
+        # Waiting for request to be finished timed out.
+        message("  Please use the ECMWF or CDS job list to track your jobs at:")
+        message("  https://apps.ecmwf.int/webmars/joblist/ or")
+        message("  https://cds.climate.copernicus.eu/cdsapp#!/yourrequests")
+        message("  and retry download using wf_transfer().")
+        message("  Note that there are user-dependent limits of submitted jobs.")
+        message("  Delete the job using wf_delete() upon completion!")
+      }
       return(ct)
     }
 
@@ -127,10 +170,14 @@ wf_request <- function(
       Sys.sleep(ct$retry)
     }
 
-    # attempt a download
-    ct <- wf_transfer(email = email,
-                      url = ct$href,
-                      verbose = verbose)
+    # attempt a download. Use 'input_user', can also
+    # be NULL (load user information from '.ecmwfapirc'
+    # file inside wf_transfer).
+    ct <- wf_transfer(user    = input_email,
+                      url      = ct$href,
+                      service  = "webapi",
+                      filename = tmp_file,
+                      verbose  = verbose)
   }
 
   # Copy data from temporary file to final location
@@ -138,24 +185,30 @@ wf_request <- function(
   # The latter to facilitate package integration.
   if (path != tempdir()) {
 
-    # create temporary output file
-    ecmwf_tmp_file <- file.path(tempdir(), "ecmwf_tmp.nc")
+    src <- file.path(tempdir(), tmp_file)
+    dst <- file.path(path, request$target)
 
-    # copy temporary file to final destination
-    file.copy(ecmwf_tmp_file,
-              file.path(path, request$target),
-              overwrite = TRUE,
-              copy.mode = FALSE)
+    if ( verbose ){
+      message(sprintf("- moved temporary file to -> %s", dst))
+    }
 
-    # cleanup of temporary file
-    invisible(file.remove(ecmwf_tmp_file))
+    # rename / move file
+    file.rename(src, dst)
+
   } else {
+    dst <- file.path(path, tmp_file)
     message("- file not copied and removed (path == tempdir())")
   }
 
   # delete the request upon succesful download
-  # to free up other download slots
-  wf_delete(email = email,
-            url = ct$href,
-            verbose = verbose)
+  # to free up other download slots. Not possible
+  # for ECMWF mars requests (skip)
+  if(!request$dataset == "mars") {
+    wf_delete(user   = input_email,
+              url     = ct$href,
+              verbose = verbose)
+  }
+
+  # return final file name/path (dst = destination).
+  return(invisible(dst))
 }