+++ R/wf_set_key.R
@@ -3,33 +3,35 @@
 #' Saves the token to your local keychain under
 #' a service called "ecmwfr".
 #'
-#' @param email email address used to sign up for the ECMWF data service
+#' @param user user (email address) used to sign up for the ECMWF data service
 #' @param key token provided by ECMWF
+#' @param service which service to use
 #' @keywords key management
 #' @seealso \code{\link[ecmwfr]{wf_get_key}}
 #' @export
+#' @author Koen Kufkens
 #' @examples
 #'
 #' \dontrun{
 #' # set key
-#' wf_set_key(email = "test@mail.com", key = "123")
+#' wf_set_key(user = "test@mail.com", key = "123")
 #'
 #' # get key
-#' wf_get_key(email = "test@mail.com")
+#' wf_get_key(user = "test@mail.com")
 #'}
 
 wf_set_key <- function(
-  email,
+  user,
   key,
   service = "webapi"
 ){
-  if(service == "webapi"){
+  if(service == webapi){
     keyring::key_set_with_value("ecmwfr_webapi",
-                       email,
+                       user,
                        password = key)
-  } else {
+  }else{
     keyring::key_set_with_value("ecmwfr_cds",
-                                email,
+                                user,
                                 password = key)
   }
 }