+++ R/wf_user_info.R
@@ -1,8 +1,8 @@
-#' ECMWF user info query
+#' ECMWF WebAPI user info query
 #'
-#' Returns user info
+#' Returns user info for the ECMWF WebAPI
 #'
-#' @param email email address used to sign up for the ECMWF data service and
+#' @param user user (email address) used to sign up for the ECMWF data service,
 #' used to retrieve the token set by \code{\link[ecmwfr]{wf_set_key}}
 #' @return returns a data frame with user info
 #' @keywords data download, climate, re-analysis
@@ -14,23 +14,23 @@
 #'
 #' \dontrun{
 #' # set key
-#' wf_set_key(email = "test@mail.com", key = "123")
+#' wf_set_key(user = "test@mail.com", key = "123")
 #'
 #' # get user info
 #' wf_user_info("test@mail.com")
 #'}
 
 wf_user_info <- function(
-  email
+  user
 ){
 
   # check the login credentials
-  if(missing(email)){
-    stop("Please provide ECMWF login email / url!")
+  if(missing(user)){
+    stop("Please provide ECMWF login user / url!")
   }
 
-  # get key from email
-  key <- wf_get_key(email)
+  # get key from user
+  key <- wf_get_key(user)
 
   # query the status url provided
   response <- httr::GET(
@@ -39,7 +39,7 @@ wf_user_info <- function(
     httr::add_headers(
       "Accept" = "application/json",
       "Content-Type" = "application/json",
-      "From" = email,
+      "From" = user,
       "X-ECMWF-KEY" = key),
     encode = "json"
   )