+++ tests/testthat/test_main_functions.r
@@ -21,12 +21,9 @@ my_request <- list(stream = "oper",
 # if provided, otherwise just continue
 # assuming a valid keychain value (see
 # additional check below)
-key <- system("echo $KEY", intern = TRUE)
-if(key != "" & key != "$KEY"){
-  wf_set_key(email = "khrdev@outlook.com",
-             key = system("echo $KEY", intern = TRUE))
-}
-rm(key)
+# Using environment variables (travis)
+wf_set_key(email = Sys.getenv("ECMWFAPIEMAIL"),
+           key   = Sys.getenv("ECMWFAPIKEY"))
 
 # Check if a password is not set. This traps the inconsistent
 # behavious between systems while accomodating for encrypted
@@ -35,8 +32,12 @@ rm(key)
 # environmental variables (hence fail to retrieve the api key).
 # This also allows for very basic checks on r-hub.
 # No checks should be skiped on either Travis CI or OSX.
-skip_check <- try(wf_get_key(email = "khrdev@outlook.com"))
-skip_check <- inherits(skip_check, "try-error")
+if(any(c(nchar(Sys.getenv("ECMWFAPIEMAIL")), nchar(Sys.getenv("ECMWFAPIKEY"))) == 0)) {
+  skip_check <- TRUE
+} else {
+  skip_check <- try(wf_get_key(email = Sys.getenv("ECMWFAPIEMAIL")))
+  skip_check <- inherits(skip_check, "try-error")
+}
 
 # check keychain management
 test_that("set, get secret key",{
@@ -48,7 +49,7 @@ test_that("set, get secret key",{
 
 test_that("test dataset function", {
   skip_if(skip_check)
-  expect_output(str(wf_datasets(email = "khrdev@outlook.com")))
+  expect_output(str(wf_datasets(email = Sys.getenv("ECMWFAPIEMAIL"))))
 })
 
 test_that("test dataset function - no login", {
@@ -57,7 +58,7 @@ test_that("test dataset function - no login", {
 
 test_that("test services function", {
   skip_if(skip_check)
-  expect_output(str(wf_services(email = "khrdev@outlook.com")))
+  expect_output(str(wf_services(email = Sys.getenv("ECMWFAPIEMAIL"))))
 })
 
 test_that("test services function - no login", {
@@ -66,7 +67,7 @@ test_that("test services function - no login", {
 
 test_that("test user info function", {
   skip_if(skip_check)
-  expect_output(str(wf_user_info(email = "khrdev@outlook.com")))
+  expect_output(str(wf_user_info(email = Sys.getenv("ECMWFAPIEMAIL"))))
 })
 
 test_that("test user info function - no login", {
@@ -76,7 +77,7 @@ test_that("test user info function - no login", {
 test_that("test request (transfer) function", {
   skip_if(skip_check)
   expect_message(wf_request(
-    email = "khrdev@outlook.com",
+    email = Sys.getenv("ECMWFAPIEMAIL"),
     transfer = TRUE,
     request = my_request,
     time_out = 60))
@@ -85,7 +86,7 @@ test_that("test request (transfer) function", {
 test_that("test request (transfer) function - time out", {
   skip_if(skip_check)
   expect_output(str(wf_request(
-    email = "khrdev@outlook.com",
+    email = Sys.getenv("ECMWFAPIEMAIL"),
     transfer = TRUE,
     request = my_request,
     time_out = 1)))
@@ -94,12 +95,12 @@ test_that("test request (transfer) function - time out", {
 test_that("test request (transfer) function - no transfer", {
   skip_if(skip_check)
   ct <- wf_request(
-    email = "khrdev@outlook.com",
+    email = Sys.getenv("ECMWFAPIEMAIL"),
     transfer = FALSE,
     request = my_request)
 
   expect_output(str(ct))
-  expect_message(wf_delete(email = "khrdev@outlook.com",
+  expect_message(wf_delete(email = Sys.getenv("ECMWFAPIEMAIL"),
                            url = ct$href))
 })
 
@@ -114,7 +115,7 @@ test_that("test transfer function - no login", {
 test_that("test request (transfer) function", {
   skip_if(skip_check)
   expect_message(wf_request(
-    email = "khrdev@outlook.com",
+    email = Sys.getenv("ECMWFAPIEMAIL"),
     transfer = TRUE,
     request = my_request,
     time_out = 180))
@@ -143,7 +144,7 @@ test_that("test request (transfer) function - larger download", {
                  target = "tmp.nc")
 
   expect_message(wf_request(
-    email = "khrdev@outlook.com",
+    email = Sys.getenv("ECMWFAPIEMAIL"),
     transfer = TRUE,
     request = large_request,
     time_out = 300))